rename filter file
