+++ src/app/search/SearchFilter.tsx
@@ -8,7 +8,7 @@ import { setSearchQuery } from '../shell/actions';
 import _ from 'lodash';
 import './search-filter.scss';
 import { destinyVersionSelector, currentAccountSelector } from '../accounts/reducer';
-import { SearchConfig, searchFilterSelector, searchConfigSelector } from './search-filters';
+import { SearchConfig, searchFilterSelector, searchConfigSelector } from './search-filter';
 import { DestinyAccount } from '../accounts/destiny-account';
 import { DimItem } from '../inventory/item-types';
 import { loadingTracker } from '../shell/loading-tracker';
